fix: [CDS-36047]: Delete helm cache dir only if both FF are enabled (#32077)

+++ 400-rest/src/main/java/software/wings/delegatetasks/helm/HelmTaskHelper.java
@@ -417,7 +417,7 @@ public class HelmTaskHelper {
           helmChartConfigParams.getChartVersion(), chartDirectory, helmChartConfigParams.getHelmVersion(),
           helmCommandFlag, timeoutInMillis, helmChartConfigParams.isCheckIncorrectChartVersion(), cacheDir);
     } finally {
-      if (helmChartConfigParams.isDeleteRepoCacheDir()) {
+      if (helmChartConfigParams.isUseRepoFlags() && helmChartConfigParams.isDeleteRepoCacheDir()) {
         try {
           FileUtils.forceDelete(new File(cacheDir));
         } catch (IOException ie) {
